TP4/requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
scipy
numpy
pyyaml
